+++ changelog.rst
@@ -1,6 +1,11 @@
 Changelog
 =========
 
+Release 0.3.24
+--------------
+
+- Fix: ``AvailableVagrantHosts``: fixed vagrant hosts parsing under Windows
+
 Release 0.3.22
 --------------